+++ apps/script/aichatCloudWatchDashboard/config.js
